refactor(filecheck): extract ./constants to /libs/{env,constants}
